PM-3851 - #4.5h start integration
