+++ mock-data/ai-review-config.json
@@ -1,48 +1,20 @@
 {
-  "id": "config_challenge_001",
-  "challengeId": "challenge_12345",
-  "minPassingThreshold": 75.00,
+  "challengeId": "97701509-f4ee-4a03-9bd1-bad7413d2274",
+  "minPassingThreshold": 15,
+  "mode": "AI_ONLY",
+  "templateId": "template_001",
   "autoFinalize": false,
-  "mode": "AI_GATING",
-  "formula": {
-    "type": "gating",
-    "gates": [
-      {
-        "workflowId": "wf_001_security",
-        "threshold": 80,
-        "isMandatory": true
-      }
-    ],
-    "scoring": {
-      "type": "weighted",
-      "weights": {
-        "wf_001_security": 0.4,
-        "wf_002_quality": 0.6
-      }
-    }
-  },
+  "formula": {},
   "workflows": [
     {
-      "id": "config_wf_001",
-      "configId": "config_challenge_001",
-      "workflowId": "wf_001_security",
-      "weightPercent": 40.00,
-      "isGating": true,
-      "createdAt": "2026-02-10T10:00:00Z",
-      "createdBy": "user_123"
+      "workflowId": "1lROGgC0jANqJL",
+      "weightPercent": 80,
+      "isGating": false
     },
     {
-      "id": "config_wf_002",
-      "configId": "config_challenge_001",
-      "workflowId": "wf_002_quality",
-      "weightPercent": 60.00,
-      "isGating": false,
-      "createdAt": "2026-02-10T10:00:00Z",
-      "createdBy": "user_123"
+      "workflowId": "J0aZLgbf9NUvnZ",
+      "weightPercent": 20,
+      "isGating": false
     }
-  ],
-  "createdAt": "2026-02-10T10:00:00Z",
-  "createdBy": "user_123",
-  "updatedAt": "2026-02-15T14:30:00Z",
-  "updatedBy": "user_123"
+  ]
 }
