+++ src/components/ChallengeEditor/ChallengeReviewer-Field/AiReviewerTab/index.js
@@ -1,0 +1,1 @@
+export { default as AiReviewTab } from './AiReviewTab';
