+++ src/components/ChallengeEditor/ChallengeReviewer-Field/AiReviewerTab/utils.js
@@ -1,0 +1,6 @@
+export const isAIReviewer = (reviewer) => {
+  return reviewer && (
+    (reviewer.aiWorkflowId && reviewer.aiWorkflowId.trim() !== '') ||
+    (reviewer.isMemberReview === false)
+  )
+}
