+++ src/config/constants.js
@@ -42,6 +42,7 @@ export const {
   PROFILE_URL,
   TC_FINANCE_API_URL,
   TC_AI_API_BASE_URL,
+  TC_REVIEWS_API_BASE_URL,
   TC_AI_SKILLS_EXTRACTION_WORKFLOW_ID,
   ENGAGEMENTS_APP_URL
 } = process.env
